requirements.txt
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
Pillow==9.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
